# backend/openapi.yaml
openapi: 3.0.3
info:
  title: App Tasks Backend V3 API
  version: 1.0.0
  description: >
    Contrato completo da API de tarefas com autenticacao JWT, persistencia em JSON
    e prefixo `/api`.
servers:
  - url: http://localhost:3000/api
    description: Ambiente local
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    ErrorResponse:
      type: object
      additionalProperties: false
      properties:
        error:
          type: string
          example: Titulo da tarefa e obrigatorio.
      required:
        - error
    User:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
          example: 78d4ed6e-d08f-4bbb-a90d-81323a019e90
        name:
          type: string
          example: Fernando
        email:
          type: string
          format: email
          example: fernando@example.com
        createdAt:
          type: string
          format: date-time
          example: 2026-03-10T21:30:00.000Z
      required:
        - id
        - name
        - email
        - createdAt
    CreateUserRequest:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 1
          example: Fernando
        email:
          type: string
          format: email
          example: fernando@example.com
        password:
          type: string
          minLength: 6
          example: 123456
      required:
        - name
        - email
        - password
    CreateUserResponse:
      type: object
      additionalProperties: false
      properties:
        user:
          $ref: '#/components/schemas/User'
      required:
        - user
...
    ProfileResponse:
      type: object
      additionalProperties: false
      properties:
        user:
          $ref: '#/components/schemas/User'
      required:
        - user
  responses:
...
    Unauthorized:
      description: Token ausente, invalido, expirado ou credenciais invalidas.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missingToken:
              value:
                error: Token nao informado.
            invalidToken:
              value:
                error: Token invalido.
            expiredToken:
              value:
                error: Token expirado.
            revokedToken:
              value:
                error: Token revogado.
            invalidCredentials:
              value:
                error: Credenciais invalidas.
...
    Conflict:
      description: Conflito de dados.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            duplicateUser:
              value:
                error: Ja existe um usuario com este email.
    InternalServerError:
      description: Erro interno inesperado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Erro interno do servidor.
paths:
...
  /users:
    post:
      tags:
        - Users
      summary: Cria um novo usuario
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
            examples:
              default:
                value:
                  name: Fernando
                  email: fernando@example.com
                  password: 123456
      responses:
        '201':
          description: Usuario criado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateUserResponse'
        '400':
          description: Erro de validacao.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missingName:
                  value:
                    error: Nome e obrigatorio.
                invalidEmail:
                  value:
                    error: Email invalido.
                missingPassword:
                  value:
                    error: Senha e obrigatoria.
                shortPassword:
                  value:
                    error: A senha deve ter pelo menos 6 caracteres.
                invalidBody:
                  value:
                    error: O corpo da requisicao deve ser um objeto JSON.
                invalidJson:
                  value:
                    error: JSON invalido.
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /users/me:
    get:
      tags:
        - Users
      summary: Retorna o perfil do usuario autenticado
      operationId: getProfile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Perfil recuperado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...